Add test utils to x/dhkem-secp256k1.
